Move existing track to separate file
