+++ src/music/constants.ts
@@ -124,32 +124,7 @@ export const resetPlaylistPrefixCommandPatterns = [
   ),
 ];
 
-export const existingTrackPattern = new RegExp(/([\d]+)/gim);
-export const playExistingTrackOptTrackPrefixCommands = ['p', 'play', 'add'];
-export const playExistingTrackMandTrackPrefixCommands = ['q', 'queue'];
-const trackPrefixTerms = ['track', 'song'];
-export const playExistingTrackPrefixCommands = [
-  /**
-   * /^;(p|play|add)( track| song)? ([\d]+)/gim,
-   */
-  [
-    `(?:${playExistingTrackOptTrackPrefixCommands.join('|')})`,
-    `(?: ${trackPrefixTerms.join('|')})?`,
-    ' ',
-    '([\\d]+)',
-  ],
-  /**
-   * Ensure that it doesn't conflict with list
-   * /^;(q|queue) (track|song) ([\d]+)/gim,
-   */
-  [
-    playExistingTrackMandTrackPrefixCommands.join('|'),
-    ' ',
-    `(?:${trackPrefixTerms.join('|')})`,
-    ' ',
-    '([\\d]+)',
-  ],
-];
+
 /**
  * @deprecated
  */
